convbib.g: improve sorting of bib items
Previously we just sorted by author name; now we also use the title of a publication
as a tie breaker if two papers have the exact same author string

Moreover, convert all strings used for sorting from Unicode to lowercase ASCII and
remove some special characters, to further reduce strangeness in the sorting result.
It's still not perfect, but a lot better than what was in place before.

Finally, we use a stable sort to get rid of any random fluctuations that might be
left in the sorting process (there shouldn't be any, but better stay on the safe side).