CI: replace Travis with GitHub Actions for building the site
This is currently quite slow (3½ minutes), but since we don't run it
very often I don't think we should spend time speeding it up just yet.

This doesn't actually do anything with the built site, but it at least
lets us see if a pull request or push is going to actually break the
website such that it doesn't build.

This runs on pushes, on PRs, and can be manually dispatched.