Script to repair cross links between manuals
This is a temporary script. It introduces symbolic links to make cross
HTML-links between manuals (referring to pkg/xxx/.. for non-existing
xxx) valid.

Once all directory names inside pkg/ are normalized and manuals are
built wrt. this structure this script is no longer needed.

Also mention this script in the README.md.