Add an imprint and privacy policy to comply with GPDR
We just link to the imprint and privacy policy of TU Kaiserslautern for
now. Since those refer explicit to a "Contact" link, also make
one clearly visible next to the other two new links in the navbar.

In exchange, remove the footer email link "The GAP Group", it seems
redundant.