Remove search page
This was implemented using Google search, which is problematic
from a data protection perspective. The whole thing was not
super helpful in my experience either, so all in all, I think
we are better off and removing it -- for now, at least, perhaps
in the future we can add a better search feature.