Remove upgrade.html
Its usefulness seems marginal at best, and it's yet another page we have
to (but do not) maintain.

This also removes a broken link to a graph showing all packages and
their dependencies (I would have removed it even if not broken, though,
as while it was (is?) a cool graph, I don't think it's useful enough to
warrant putting it in that prominent place).