Tweak a pointer to the support email address
Moderators of the support mailing list are seeing a LOT of emails coming
from people who really want to contact the support of the GAP clothing
company. They are finding our email address via a Google search for 'how
do I email gap', which prompts Google to show a 'snippet' text from our
website -- namely the text modified by this patch. With the revision, I
hope that either Google will stop showing this snippet, or that at least
some people will realize that this is the wrong GAP.