Centralize package status into a single file
Instead of letting each package keep track of its own status as a
deposited or refereed package, we now collect this information
centrally. This makes more sense, and also avoids some awkwardness

For example, in order to be accepted for distribution, currently GAP
package authors must provide a proper release of their package, which
then is reviewed. But that release should not yet have status `accepted`
resp. `deposited`... so strictly speaking, right after the package is
officially accepted, a second release with the updated status should be
made right away.

With a centralized list of deposited and accepted packages, there is no
need anymore for packages metadata to contain this information.