Alternative convbib_mj.g for convbib.g
The file convbib_mj.g produces a MathJax version of gap-published.html.

This is done by converting to BibXMLext and producing the HTML
from there (this has a MathJax option).

There are quite a few entries which are not valid BibTeX and the
conversion to BibXMLext will not work. In those cases we use the
simpler method from convbib.g as a fallback.

There is math code in a few bib-entries which is not supported
by MathJax.